add me and auth api
