+++ pkg/utils/helper.go
@@ -1,0 +1,8 @@
+package utils
+
+type UserInfo struct {
+	Type           string `json:"type"`
+	UserId         string `json:"user_id"`
+	Name           string `json:"name"`
+	IdentityNumber string `json:"identity_number"`
+}
